Automatic route mounting configuration
Automatically mount the engine, by default, on `/console`. Do so only in
the `development` and `test` environments. `test` is in there, just so
we can test it without hacking the app in the tests.

The default mount path of `/console` can be changed from
`config.web_console.default_mount_path`.