Adapter registration supports a block
For the IRB adapter we had to do monkey patching and rails console
helper methods inclusion. Both of those were a one-off tasks, that were
placed kinda funky in the code.

If WebConsole::REPL.register_adapter supports a block, that would be
executed right after the initialization, those blocks would be prefect
locations for one-off tasks, like the one above.