Dummy adapter implementation
The dummy adapter is there as a fallback when the currently selected
REPL console in rails is unsupported. It does not wraps around an
existing REPL, it just evaluates code in a binding and takes care of the
error reporting.

There is no multiline support, yet. I'm gonna keep it in the back of my
mind as a low-priority thing.