Use term.js to drop GPL code
GPL is kinda like cancer, it spreads. While I did have contacted the
author of shellinabox, I still don't want the possibility of the need to
relicense the whole code base into GPL.

So as of now, I would like to change the terminal emulator to @chjj's
`term.js`. It started out as a fork of Fabrice Bellard's javascript
vt100 for jslinux, it has been re-licensed to MIT, with the approval of
Fabrice Bellard.

I tried to styled it much as the shellinabox's vt100.js, so the end user
would probably not notice the difference.