Bring back 1.0 features without automount
We removed the code, so its easier for Ryan to develop, with no legacy
code to think or worry about. Let's bring it back for the final 2.0
release.

It carries a twist though, I'm disabling the default `/console` mount.
If the users still want the old functionally, they have to explicitly
enable it.