Push the error page body upwards the size of the console.
While, I wouldn't like to do that on every custom page (so I don't screw
user's layouts), I think a lot of developers want to see all of the content
on the default Rails error page.

Since it's quite special as is now, being a bit more special in the name of
better user experience, won't hurt.