Fix templates non rendering, because of missing template suffix
This seems to regressed on the latest ActionView, which doesn't seem to
render plain `html` files with ERB expansions.

I initially dropped the `.erb` part with the hope that more designers
can contribute to the project, having it nicely highlighted in an
arbitrary editor.

The reason, I didn't catch this early was because the master Gemfile
wasn't bundling on the CI and I'm trying to fix that as well, so I can
prevent other regressions in future.