Remove JRuby support for 3.0
I want to revisit this, but I have to extract the hacky bindex
implementation for JRuby 9k. Will do this for 3.1 since we never saw
that much JRuby usage for the effort we put into supporting it.

I still want to support it, it's just not a priority at the moment.