Log internal web-console errors
Currently, if an error is hit in web-console internals, the user see's
the `ActionPack::ShowExceptions` middleware page and nothing else.
Having a log of the error within the Rails application log can help the
developers send us bug reports or take an action in resolving it.

Thanks to Richard Schneeman for the push and implementation.