Let #console live in Kernel
With this change, we can render a console from any part of an Rails
application, not just in the controllers or the views. Wanna introspect
a model or your plain old Ruby class? No problem!

Hopefully, this change can boost up the usage of the `#console` helper.

To do this juggling we now persist a binding or an exception in a fiber
local variable. The middleware can create a console session from those,
when needed. This is a global state, so we take precautions to clean it
between requests, even in multi-threaded environments.