Let FakeMiddleware to load "lib/web_console.rb"
On testing templates, the FakeMiddleware caused an error when the "WebConsole.looger"
is missing, so we couldn't run tests for templates.

Previously, the "lib/web_console.rb" required all modules, and it caused some troubles on
the middleware, but now it is using ActiveSupport::Autoload, so we don't worry about the
troubles and the middleware can require it.