Map bindings to traces based on the trace __FILE__ and __LINE__
Currently, we naively map exception bindings by the exception
backtrace's indexes. This doesn't work as the backtraces and the
bindings doesn't map one to one and we're most likely to be out of sync
in a couple of trace indexes.

With this change, I try to be smarter about the binding that corresponds
to an exception backtrace index. The `ExceptionMapper` objects tries to
map to a binding based on the __FILE__ and __LINE__ of the trace behind
that index. It also hides this transformation in an interface similar to
an `Array`, so we didn't needed to change much backend code and any
frontend code.

Right now, if you go and click through all the traces, you get the
binding in most of the times. I can imagine we can still be off in
exceptions with custom backtraces, but we'll be off only for a couple of
bindings and not everything after those special traces.

In any case, I think this can fix a lot of the `web-console doesn't
work` cases for Rails 5.