Let WebConsole.logger respect Rails.logger
We used to set WebConsole.logger during Bundle.require time, which
happened to be way earlier than when Rails.logger was initialized, so we
always ended up with `STDERR` logger. In this change, we let
`WebConsole.logger` to dynamically call `Rails.logger`, which will let
us use it once initialized. If by the time we try to log something,
`Rails.logger` isn't initialized or is explicitly set to nothing, we'll
still fallback to the `STDERR` logging.